requirements.txt
Django==1.9
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.3.3
djangorestframework-gis==0.10.1
GDAL==1.11.2
psycopg2==2.6.2
python-dateutil==2.5.3
requests==2.9.1
